First step in separation into modules
Does not work yet. WIP!

Refers issue to #7
